Read an integer argument rather than casting one
Four of Copilot's five findings, and the interesting one is the fix not
taken.

**An integer parameter is now read checked.** `static_cast<int32_t>(
in.GetF64())` is undefined for a NaN, an infinity or a 1e300 — exactly what
`GetId` was before `IsRepresentableId`, one door along. `wire::GetInteger<T>`
refuses a value the conversion is not defined for and fails the reader, so
the facade answers "malformed arguments" instead of passing on whatever the
cast produced. Bounds are exact: for an integer of four bytes or fewer every
value is a double, and anything wider is refused at compile time rather than
checked approximately, because 2^63 is not a double and a bound that rounds
the wrong way passes the one value it exists to stop.

**The wire format is deliberately untouched.** The review asked for the wire
kind to come from the C++ type, so `int32_t` would cross as `i32`. Declined
for parameters alone: every `int32_t` *field* in the codec has crossed as an
f64 since ADR 0013, and making one C++ type cross one way as a parameter and
another as a field is worse than the lossy-but-uniform rule. Moving both
together is a change of its own with a golden-hex update in both suites, and
it stays filed in ADR 0038's consequences — now with the half that is done
separated from the half that is not.

**A qualified integer parameter is still that integer.** The parser keeps a
type as written and the subset allows `const` and `&`, so `const int32_t`
missed the table and fell back to `double` — reinstating the narrowing in
the one spelling nobody would test again.

**The engine's accumulators are 64-bit.** With `maxEdge` at 1 a block is the
whole frame, so a channel's sum is the pixel count times 255 and a frame past
about 4100 a side overflowed a `uint32_t` into an average darker than
anything in the picture. Unreachable today at a 1280 cap, and no test asserts
it: the frame that would demonstrate it is 67 MB, half a phone's whole store
ceiling. Widening costs nothing and takes the trap out of the arithmetic.

**The candidate search stops at the first match**, which is what a unique
identity means.

Tests first for the two that are testable. Six values that are not integers
and two that are exactly the bounds, so a check written with the wrong
comparison rejects the largest valid value; three qualified spellings; and
both halves of "an angle is not a count".

Sabotages predicted first. Dropping the canonicalisation fails exactly the
qualified test. Reverting to the cast fails the checked test and the
generated-files staleness check — the second is the right extra signal for a
generator change, and I had not predicted it.

+++ core/src/engines/frame_preview_engine/box_frame_preview_engine.cpp
@@ -98,10 +98,17 @@ Result<FramePreview> BoxFramePreviewEngine::Reduce(const FrameRef& frame, int32_
   preview.height = static_cast<int32_t>(rows);
   preview.pixels.assign(static_cast<size_t>(cols * rows * 4), 0u);
 
-  const uint32_t area = static_cast<uint32_t>(block * block);
+  // Sixty-four bits, for the same reason the offsets above are. A block is the whole frame when
+  // `maxEdge` is 1, so `area` is the pixel count and each channel's sum is that times 255: a
+  // frame past about 4100 on a side overflows a `uint32_t` and the average comes out darker than
+  // anything in the picture. No frame here is that large yet — the grabber caps its long edge at
+  // 1280 — and no test asserts it, because the frame that would demonstrate it is 67 MB, which is
+  // half a phone's whole frame-store ceiling (ADR 0023). Widening it costs nothing and removes
+  // the trap from the arithmetic rather than from a comment.
+  const uint64_t area = static_cast<uint64_t>(block) * static_cast<uint64_t>(block);
   for (int64_t row = 0; row < rows; ++row) {
     for (int64_t col = 0; col < cols; ++col) {
-      std::array<uint32_t, 3> sum{0, 0, 0};
+      std::array<uint64_t, 3> sum{0, 0, 0};
       for (int64_t dy = 0; dy < block; ++dy) {
         for (int64_t dx = 0; dx < block; ++dx) {
           const auto rgb = ColourAt(pinned.value, frame.format, stride, col * block + dx,
